SRSoD/pyproject.toml
[tool.poetry]
name = "scooter-repair-service"
version = "0.1.0"
description = "a web used to book scooter servive"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
authlib = "^1.3.2"
httpx = "^0.27.2"
starlette-session = "^0.4.3"
python-dotenv = "^1.0.1"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
